+++ OpenKh.Command.IdxImg/OpenKh.Command.IdxImg.csproj
@@ -16,6 +16,7 @@
     <ProjectReference Include="..\OpenKh.Egs\OpenKh.Egs.csproj" />
     <ProjectReference Include="..\OpenKh.Kh1\OpenKh.Kh1.csproj" />
     <ProjectReference Include="..\OpenKh.Kh2\OpenKh.Kh2.csproj" />
+    <ProjectReference Include="..\OpenKh.Patcher\OpenKh.Patcher.csproj" />
   </ItemGroup>
 
 </Project>
