--- OpenKh.Command.Patcher/OpenKh.Command.Patcher.csproj
@@ -1,15 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>net6.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\OpenKh.Common\OpenKh.Common.csproj" />
-    <ProjectReference Include="..\OpenKh.Patcher\OpenKh.Patcher.csproj" />
-  </ItemGroup>
-
-</Project>
