+++ OpenKh.sln
@@ -214,8 +214,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Simple3DViewport", "Simple3
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OpenKh.Tools.Kh2ObjectEditor", "OpenKh.Tools.Kh2ObjectEditor\OpenKh.Tools.Kh2ObjectEditor.csproj", "{CF6D785D-602A-46E0-9DCB-5F703E3544F0}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OpenKh.Command.Patcher", "OpenKh.Command.Patcher\OpenKh.Command.Patcher.csproj", "{29512ADC-EAD0-4260-AE30-5B76CD0C66A0}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		.NET Debug|Any CPU = .NET Debug|Any CPU
@@ -1504,22 +1502,6 @@ Global
 		{CF6D785D-602A-46E0-9DCB-5F703E3544F0}.Release|Any CPU.Build.0 = Release|Any CPU
 		{CF6D785D-602A-46E0-9DCB-5F703E3544F0}.Release|x64.ActiveCfg = Release|Any CPU
 		{CF6D785D-602A-46E0-9DCB-5F703E3544F0}.Release|x64.Build.0 = Release|Any CPU
-		{29512ADC-EAD0-4260-AE30-5B76CD0C66A0}..NET Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{29512ADC-EAD0-4260-AE30-5B76CD0C66A0}..NET Debug|Any CPU.Build.0 = Debug|Any CPU
-		{29512ADC-EAD0-4260-AE30-5B76CD0C66A0}..NET Debug|x64.ActiveCfg = Debug|Any CPU
-		{29512ADC-EAD0-4260-AE30-5B76CD0C66A0}..NET Debug|x64.Build.0 = Debug|Any CPU
-		{29512ADC-EAD0-4260-AE30-5B76CD0C66A0}..NET Release|Any CPU.ActiveCfg = Release|Any CPU
-		{29512ADC-EAD0-4260-AE30-5B76CD0C66A0}..NET Release|Any CPU.Build.0 = Release|Any CPU
-		{29512ADC-EAD0-4260-AE30-5B76CD0C66A0}..NET Release|x64.ActiveCfg = Release|Any CPU
-		{29512ADC-EAD0-4260-AE30-5B76CD0C66A0}..NET Release|x64.Build.0 = Release|Any CPU
-		{29512ADC-EAD0-4260-AE30-5B76CD0C66A0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{29512ADC-EAD0-4260-AE30-5B76CD0C66A0}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{29512ADC-EAD0-4260-AE30-5B76CD0C66A0}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{29512ADC-EAD0-4260-AE30-5B76CD0C66A0}.Debug|x64.Build.0 = Debug|Any CPU
-		{29512ADC-EAD0-4260-AE30-5B76CD0C66A0}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{29512ADC-EAD0-4260-AE30-5B76CD0C66A0}.Release|Any CPU.Build.0 = Release|Any CPU
-		{29512ADC-EAD0-4260-AE30-5B76CD0C66A0}.Release|x64.ActiveCfg = Release|Any CPU
-		{29512ADC-EAD0-4260-AE30-5B76CD0C66A0}.Release|x64.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -1609,7 +1591,6 @@ Global
 		{1D074EA4-9964-43EC-9389-7E60C2545C13} = {0FB7CD6A-EE31-467D-A590-267DCD028618}
 		{EAD6D51F-5106-4E11-A4E4-104A18C0235B} = {9C52D787-9819-4B89-989B-EEA6FEB20731}
 		{CF6D785D-602A-46E0-9DCB-5F703E3544F0} = {402B2669-D594-4DFA-965B-02A92626ADC6}
-		{29512ADC-EAD0-4260-AE30-5B76CD0C66A0} = {9C52D787-9819-4B89-989B-EEA6FEB20731}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {1DF3960B-4FE5-4E56-92D4-2FC0A912E823}
